Change exponential convolutions to use recursive_filter and moved to pole_zero.py
