+++ src/dspeed/processors/__init__.py
@@ -66,8 +66,6 @@ from importlib import import_module
 _modules = {
     "mean_below_threshold": "arithmetic",
     "bl_subtract": "bl_subtract",
-    "convolve_damped_oscillator": "convolutions",
-    "convolve_exp": "convolutions",
     "convolve_wf": "convolutions",
     "fft_convolve_wf": "convolutions",
     "reflected_convolve_wf": "convolutions",
@@ -91,7 +89,6 @@ _modules = {
     "iir_filter": "iir_filter",
     "notch_filter": "iir_filter",
     "peak_filter": "iir_filter",
-    "inject_damped_oscillation": "inject_ringing",
     "inl_correction": "inl_correction",
     "moving_slope": "kernels",
     "step": "kernels",
@@ -123,6 +120,9 @@ _modules = {
     "inject_gumbel": "pmt_pulse_injector",
     "pole_zero": "pole_zero",
     "double_pole_zero": "pole_zero",
+    "convolve_damped_oscillator": "pole_zero",
+    "convolve_exp": "pole_zero",
+    "inject_damped_oscillation": "pole_zero",
     "poly_diff": "poly_fit",
     "poly_exp_rms": "poly_fit",
     "poly_fit": "poly_fit",
