+++ src/dspeed/processors/recursive_filter.py
@@ -11,7 +11,7 @@ from ..utils import numba_defaults_kwargs as nb_kwargs
 
 @guvectorize(
     [
-        "void(float32[:], float32[:], float32[:], float32, float32, float32[:])",
+        "void(float32[:], float64[:], float64[:], float32, float32, float32[:])",
         "void(float64[:], float64[:], float64[:], float64, float64, float64[:])",
     ],
     "(n),(p),(q),(),()->(n)",
